Simplified the creation of TsFloatList
I decided to always return a new instance of a list instead to keep a single instance. This simplifies the code in Python and C#. Keeping a single instance of the lists in Java is feasible but I would like to stay consistent in the bindings.

Furthermore, this patch contains a bug fix for the ts_toString method which crashed for empty lists.