Use ts_bspline_clean on the spline created by ts_bspline_interpolate_cubic() before calling ts_bspline_derive().
See https://github.com/msteinbeck/tinyspline/issues/99

A simple solution is to reduce to multiplicity of an internal knot by 1
if it is equals to order (deg + 1) and if the two corresponding control points
(https://github.com/msteinbeck/tinyspline/blob/master/library/tinyspline.h#L240)
are equals
 (you can use: https://github.com/msteinbeck/tinyspline/blob/master/library/tinyspline.h#L889
to check if two control points are "equals").
When reducing the multiplicity of such a knot we have to merge the two control points by
shifting the control point array to left so that the second control point corresponding
to our knot is replaced by the next one and so forth:
// control points
[p0, p1, p2, *p3, p4, p5, p6] // p2 and p3 need to be merged
[p0, p1, p2, p4, p5, p6] // fixed
Similarly, we can shift the knot array:
[0, 0, 0, 0, 1, 2, 3, 4, 4, 4, *4, 5, 6, 7, 8, 9, 9, 9, 9] // 4 needs to be merged
[0, 0, 0, 0, 1, 2, 3, 4, 4, 4, 5, 6, 7, 8, 9, 9, 9, 9] // fixed
This step can be done before performing the actual derivation.